Removed Decision Tree for omega_tor & Custom File Name
Commented out lines 146-173 for the decision tree to decide if omega_tor should be omega_tor_Er OR omega_tor_Vor. This was based on the decision that omega_tor_Er is the only quantity necessary to calculate the shear in the pedestal region.

Made the naming of the output iterdb file decided by user input and printed out filepath, g-file, and p-file names to help in naming. Lines 204-211

Added quick input description on Line 1 for what inputs are required and proper formatting to run script.